Initial attempt at separating actions
